+++ R/rules-spaces.R
@@ -18,7 +18,8 @@ set_space_around_op <- function(pd_flat, strict) {
   }
   if (sum(pd_flat$lag_newlines) > 2 &&
     is_function_call(pd_flat) &&
-    any(pd_flat$token %in% c("EQ_SUB", "','"))
+    any(pd_flat$token %in% c("EQ_SUB", "','")) &&
+    !getOption("styler.ignore_alignment", FALSE)
   ) {
     is_on_aligned_line <- token_is_on_aligned_line(pd_flat)
   } else {