+++ NEWS.md
@@ -1,3 +1,7 @@
+* new R option `styler.ignore_alignment` controls if alignment should be 
+  detected (and preserved) or not (#932).
+* the cache is also invalidated on changing the stylerignore markers (#932).
+
 # styler 1.7.0
 
 * if `else` follows directly after `if`, line breaks are removed (#935).